frontend/requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
python-dotenv==1.0.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
apify-client==1.6.1
langchain==0.1.12
langchain-openai==0.0.8
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.18.1
webdriver-manager==4.0.1
openpyxl==3.1.2
schedule==1.2.1 